pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentmail-proxy"
version = "0.3.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "aioredis==2.0.1",
    "mail-parser==3.15.0",
    "beautifulsoup4==4.12.2",
    "html2text==2020.1.16",
    "aiosmtplib==3.0.1",
    "aioimaplib==1.0.1",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-benchmark==4.0.0",
    "pytest-cov==4.1.0",
    "pytest-xdist==3.3.1",
    "python-dotenv==1.0.0",
    "structlog==23.2.0",
    "httpx==0.25.2"
]

[project.optional-dependencies]
dev = ["black", "ruff"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
